+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   inside a container to the inspect report. CNJR-11978
 - Etcd cluster members check that captures the current cluster members
   (as reported by `evoke cluster member list`) in the inspect report. CNJR-11979
+- Ruby thread dump check that captures thread dumps from Ruby processes running
+  in the container. CNJR-8736
 
 ## [0.4.2] - 2025-03-25